Move CtrlConnection impl to class
